feat: Allowed displaying multiple members within 1 group
