+++ components/adminComponents/userApplicationAdmin/UserAdminView.tsx
@@ -1,26 +1,10 @@
 import { useEffect, useRef, useState } from 'react';
 import { RequestHelper } from '../../../lib/request-helper';
-import Pagination from './UserAdminPagination';
 import { useAuthContext } from '../../../lib/user/AuthContext';
-import {
-  CheckIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  LockClosedIcon,
-  LockOpenIcon,
-  XIcon,
-} from '@heroicons/react/solid';
-import Link from 'next/link';
-import { getGroupId } from './helpers';
+import { LockClosedIcon, LockOpenIcon, XIcon } from '@heroicons/react/solid';
 
 interface UserAdminViewProps {
-  userGroups: UserIdentifier[][];
-  currentUserGroupId: string;
-  goBack: () => void;
-  // updateCurrentUser: (value: Omit<UserIdentifier, 'scans'>) => void;
-  onUserGroupClick: (id: string) => void;
-  // onAcceptReject: (status: string, notes: string) => void;
-  // onUpdateRole: (newRole: UserPermission) => void;
+  currentApplicant: UserIdentifier;
 }
 
 interface BasicInfoProps {
@@ -64,76 +48,16 @@ function FRQInfo({ k, v }: BasicInfoProps) {
   );
 }
 
-export default function UserAdminView({
-  // onAcceptReject,
-  // onUpdateRole,
-  userGroups,
-  currentUserGroupId,
-  goBack,
-  onUserGroupClick,
-}: UserAdminViewProps) {
+export default function UserAdminView({ currentApplicant }: UserAdminViewProps) {
   const { user } = useAuthContext();
 
-  let currentUserGroupIndex = 0;
-  const currentUserGroup = userGroups.find((group, i) => {
-    if (getGroupId(group) === currentUserGroupId) {
-      currentUserGroupIndex = i;
-      return true;
-    }
-    return false;
-  });
-
-  const stringifyScore = (appScore: { acceptCount: number; rejectCount: number }) => {
-    if (appScore.acceptCount >= 1000000000) return 'Auto-Accepted by HackPortal';
-    return `${appScore.acceptCount - appScore.rejectCount} (${appScore.acceptCount} accepted, ${
-      appScore.rejectCount
-    } rejected)`;
-  };
-
-  const user_info = [
-    ['Major', currentUserGroup[0].major],
-    ['Application Score', stringifyScore(currentUserGroup[0].applicationScore)],
-    ['University', currentUserGroup[0].university],
-    ['Current Level of Study', currentUserGroup[0].studyLevel],
-    ['Number of Hackathons Attended', currentUserGroup[0].hackathonExperience],
-    ['Software Experience', currentUserGroup[0].softwareExperience],
-    [
-      'Resume',
-      !currentUserGroup[0].resume || currentUserGroup[0].resume === '' ? (
-        <p>No resume found</p>
-      ) : (
-        <Link passHref href={currentUserGroup[0].resume} target="_blank" rel="noopener noreferrer">
-          <button className="border-2 p-3 hover:bg-gray-200">Click here to download resume</button>
-        </Link>
-      ),
-    ],
-  ];
-
   // Pagination
-  const ref = useRef(null);
   const applicationNotesRef = useRef<HTMLTextAreaElement>(null);
 
-  const [windowHeight, setWindowHeight] = useState(window.innerHeight);
-  const [height, setHeight] = useState(60);
-  const [currentPage, setCurrentPage] = useState(1);
   const [newRole, setNewRole] = useState('');
   const [errors, setErrors] = useState<string[]>([]);
-  const [isInEditMode, setIsInEditMode] = useState(false);
-
   // Contains info of the user who is viewing the data
   const { user: organizer } = useAuthContext();
-
-  useEffect(() => {
-    const handleResize = () => setWindowHeight(window.innerHeight);
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-
-  useEffect(() => {
-    const h = Math.max(60, ref.current.offsetHeight);
-    setHeight(h);
-    setCurrentPage(Math.floor(currentUserGroupIndex / Math.floor(h / 60) + 1));
-  }, [windowHeight, currentUserGroupIndex]);
 
   const updateRole = async () => {
     if (!organizer.permissions.includes('super_admin')) {
@@ -156,7 +80,7 @@ export default function UserAdminView({
           },
         },
         {
-          userId: currentUserGroup[0].id,
+          userId: currentApplicant.id,
           newRole,
         },
       );
@@ -164,281 +88,165 @@ export default function UserAdminView({
         setErrors([...errors, data.msg]);
       } else {
         alert(data.msg);
-        // onUpdateRole(newRole as UserPermission);
       }
     } catch (error) {
       console.error(error);
       setErrors((prev) => [...prev, 'Unexpected error. Please try again later']);
     }
-
-    // TODO: Make request to backend to update user roles
   };
-
-  //////////////// TODO: IMPLEMENT
-  const openResume = () => {};
-
-  const pageSize = Math.floor(height / 60);
-  const startIndex = (currentPage - 1) * pageSize;
 
   // 208 px
   return (
-    <div
-      className={`
-        lg:mx-14 flex flex-row h-full 
-        bg-[rgba(255,255,255,0.6)] backdrop-blur
-        rounded-2xl
-      `}
-    >
-      {/* User List */}
-      <div className="hidden md:block md:w-72 px-2 py-4">
-        {/* Page */}
-        <div className="overflow-y-hidden h-[calc(100%-40px)]" ref={ref}>
-          {userGroups.slice(startIndex, startIndex + pageSize).map((group) => (
-            <div
-              key={getGroupId(group)}
-              className={`
-                flex flex-row justify-center items-center w-full py-2 rounded-xl mb-3 h-12 p-4
-                bg-[rgba(255,255,255,0.6)]
-                shadow-md ${
-                  getGroupId(group) === currentUserGroupId ? 'border-primaryDark border-[2px]' : ''
-                }
-                cursor-pointer
-              `}
-              onClick={() => onUserGroupClick(getGroupId(group))}
-            >
-              {/* <div
-                className={`
-                  text-[rgba(19,19,19,1)] font-bold
-                  whitespace-nowrap overflow-hidden text-ellipsis max-w-[50%]
-                `}
-              >
-                {user.user.firstName}
-              </div> */}
-
-              <div
-                className={`
-                  py-1 px-6 text-sm font-bold rounded-full
-                  flex-1 flex flex-row justify-center items-center
-                  whitespace-nowrap overflow-hidden text-ellipsis
-                  ${
-                    group[0].status === 'Accepted'
-                      ? 'bg-[rgb(242,253,226)] text-[rgb(27,111,19)]'
-                      : ''
-                  }
-                  ${
-                    group[0].status === 'Rejected'
-                      ? 'bg-[rgb(255,233,218)] text-[rgb(122,15,39)]'
-                      : ''
-                  }
-                  ${
-                    group[0].status === 'In Review'
-                      ? 'bg-[rgb(213,244,255)] text-[rgb(9,45,122)]'
-                      : ''
-                  }
-                `}
-              >
-                {group[0].status}
-              </div>
-            </div>
-          ))}
-        </div>
-
-        {/* Pagination */}
-        <Pagination
-          currentPage={currentPage}
-          totalCount={userGroups.length}
-          pageSize={pageSize}
-          onPageChange={(page) => setCurrentPage(page)}
-        />
-      </div>
-
-      {/* User */}
-      <div className="rounded-2xl h-full overflow-y-scroll w-full flex-1">
-        {/* Header */}
-        <div className="sticky top-0 bg-[rgba(255,255,255,0.8)] flex flex-row justify-between items-center text-secondary">
-          <div className="flex items-center p-0">
-            <ChevronLeftIcon
-              className="h-12 w-12 text-sm font-extralight cursor-pointer"
-              onClick={() =>
-                onUserGroupClick(
-                  userGroups[currentUserGroupIndex - 1]
-                    ? getGroupId(userGroups[currentUserGroupIndex - 1])
-                    : '',
-                )
-              }
-            />
-            <ChevronRightIcon
-              className="h-12 w-12 text-sm font-extralight cursor-pointer"
-              onClick={() =>
-                onUserGroupClick(
-                  userGroups[currentUserGroupIndex + 1]
-                    ? getGroupId(userGroups[currentUserGroupIndex - 1])
-                    : '',
-                )
-              }
-            />
-          </div>
-          <div onClick={goBack} className="p-3">
-            <XIcon className="h-12 w-12 cursor-pointer" />
-          </div>
-        </div>
-
-        {/* Application */}
-        <div className="p-5 sm:p-10 text-complementary bg-[rgba(255,255,255,0.4)]">
-          {/* Application Status */}
-          <div className="flex-wrap gap-y-2 flex flex-row justify-between items-center">
-            <p
-              className={`
+    <div className="p-5 sm:p-10 text-complementary bg-[rgba(255,255,255,0.4)] mb-4">
+      {/* Application Status */}
+      <div className="flex-wrap gap-y-2 flex flex-row justify-between items-center">
+        <p
+          className={`
                 font-bold py-1 px-6 rounded-full
                 ${
-                  currentUserGroup[0].status === 'Accepted'
+                  currentApplicant.status === 'Accepted'
                     ? 'bg-[rgb(242,253,226)] text-[rgb(27,111,19)]'
                     : ''
                 }
                 ${
-                  currentUserGroup[0].status === 'Rejected'
+                  currentApplicant.status === 'Rejected'
                     ? 'bg-[rgb(255,233,218)] text-[rgb(122,15,39)]'
                     : ''
                 }
                 ${
-                  currentUserGroup[0].status === 'In Review'
+                  currentApplicant.status === 'In Review'
                     ? 'bg-[rgb(213,244,255)] text-[rgb(9,45,122)]'
                     : ''
                 }
               `}
-            >
-              {currentUserGroup[0].status}
-            </p>
+        >
+          {currentApplicant.status}
+        </p>
 
-            <div className="text-sm flex-wrap gap-y-2 flex flex-row justify-between items-start gap-x-3">
-              <button
-                className="rounded-full bg-transparent text-[rgba(66,184,187,1)] border-2 border-solid border-[rgba(66,184,187,1)] font-bold py-2 px-8 hover:border-red-500 hover:text-white hover:bg-red-500 transition"
-                // onClick={() => onAcceptReject('Rejected', applicationNotesRef.current.value)}
-              >
-                REJECT
-              </button>
-              <button
-                className="rounded-full bg-[rgba(66,184,187,1)] text-white border-2 border-solid border-[rgba(66,184,187,1)] font-bold py-2 px-8 hover:border-green-500 hover:bg-green-500 transition"
-                // onClick={() => onAcceptReject('Accepted', applicationNotesRef.current.value)}
-              >
-                ACCEPT
-              </button>
-            </div>
+        <div className="text-sm flex-wrap gap-y-2 flex flex-row justify-between items-start gap-x-3">
+          <button
+            className="rounded-full bg-transparent text-[rgba(66,184,187,1)] border-2 border-solid border-[rgba(66,184,187,1)] font-bold py-2 px-8 hover:border-red-500 hover:text-white hover:bg-red-500 transition"
+            // onClick={() => onAcceptReject('Rejected', applicationNotesRef.current.value)}
+          >
+            REJECT
+          </button>
+          <button
+            className="rounded-full bg-[rgba(66,184,187,1)] text-white border-2 border-solid border-[rgba(66,184,187,1)] font-bold py-2 px-8 hover:border-green-500 hover:bg-green-500 transition"
+            // onClick={() => onAcceptReject('Accepted', applicationNotesRef.current.value)}
+          >
+            ACCEPT
+          </button>
+        </div>
+      </div>
+
+      <div className="my-6 w-full border-2 border-gray-200 rounded-md" />
+
+      {/* Basic Information */}
+      <div className="flex flex-col-reverse sm:flex-row gap-2">
+        {/* Name, School, etc... */}
+        <div className="flex flex-row gap-6 sm:w-3/5">
+          <div className="flex flex-col basis-0 flex-grow">
+            <BasicInfo
+              k="Name"
+              v={currentApplicant.user.firstName + ' ' + currentApplicant.user.lastName}
+              locked={true}
+              canUnlock={organizer.permissions.includes('super_admin')}
+            />
+            <BasicInfo k="Major" v={currentApplicant.major} />
+            <BasicInfo k="Level of Study" v={currentApplicant.studyLevel} />
           </div>
 
-          <div className="my-6 w-full border-2 border-gray-200 rounded-md" />
-
-          {/* Basic Information */}
-          <div className="flex flex-col-reverse sm:flex-row gap-2">
-            {/* Name, School, etc... */}
-            <div className="flex flex-row gap-6 sm:w-3/5">
-              <div className="flex flex-col basis-0 flex-grow">
-                <BasicInfo
-                  k="Name"
-                  v={currentUserGroup[0].user.firstName + ' ' + currentUserGroup[0].user.lastName}
-                  locked={true}
-                  canUnlock={organizer.permissions.includes('super_admin')}
-                />
-                <BasicInfo k="Major" v={currentUserGroup[0].major} />
-                <BasicInfo k="Level of Study" v={currentUserGroup[0].studyLevel} />
-              </div>
-
-              <div className="flex flex-col basis-0 flex-grow">
-                <BasicInfo
-                  k="School"
-                  v={currentUserGroup[0].university ?? currentUserGroup[0].universityManual}
-                  locked={true}
-                  canUnlock={organizer.permissions.includes('super_admin')}
-                />
-                <BasicInfo k="Software Experience" v={currentUserGroup[0].softwareExperience} />
-                {/* this should PROBABLY be currentUser.hackathonNumber, but I think the registration interface is populated incorrectly */}
-                <BasicInfo
-                  k="Hackathons Attended"
-                  v={`${currentUserGroup[0].hackathonExperience}`}
-                />
-              </div>
-            </div>
-
-            {/* Application Score */}
-            <div className="flex flex-col items-center sm:w-2/5 text-black">
-              <p className="font-bold text-xl text-black">Application Score</p>
-
-              <p className="text-8xl font-dmSans">
-                {currentUserGroup[0].applicationScore.acceptCount -
-                  currentUserGroup[0].applicationScore.rejectCount}
-              </p>
-
-              <p className="italic text-gray-600">
-                <span className="text-green-500">
-                  {currentUserGroup[0].applicationScore.acceptCount} accepted
-                </span>{' '}
-                /{' '}
-                <span className="text-red-500">
-                  {currentUserGroup[0].applicationScore.rejectCount} rejected
-                </span>
-              </p>
-
-              {user.permissions.includes('super_admin') && (
-                <div className="flex-wrap flex flex-row gap-2 p-3 w-full">
-                  <select
-                    value={newRole}
-                    onChange={(e) => {
-                      setNewRole(e.target.value);
-                    }}
-                    name="new_role"
-                    className="flex-1 border-2 rounded-xl p-2"
-                  >
-                    <option value="">Choose a role</option>
-                    <option value="super_admin">Super Admin</option>
-                    <option value="admin">Admin</option>
-                    <option value="hacker">Hacker</option>
-                  </select>
-
-                  {newRole !== '' && (
-                    <button
-                      onClick={() => {
-                        updateRole();
-                      }}
-                      className="flex-1 font-bold bg-[rgba(66,184,187,1)] text-white p-2 rounded-xl"
-                    >
-                      Update Role
-                    </button>
-                  )}
-                </div>
-              )}
-            </div>
-          </div>
-
-          {/* Notes */}
-          <div>
-            <textarea
-              ref={applicationNotesRef}
-              className="w-full h-52 border-gray-300 border-2 rounded-md text-black bg-gray-300/20"
-            ></textarea>
-          </div>
-          <div className="my-6 w-full border-2 border-gray-200 rounded-md" />
-          {/* FRQ */}
-          <div className="flex flex-col w-full">
-            <FRQInfo
-              k={'Why do you want to attend HackUTD Ripple Effect?'}
-              v={currentUserGroup[0].whyAttend}
+          <div className="flex flex-col basis-0 flex-grow">
+            <BasicInfo
+              k="School"
+              v={currentApplicant.university ?? currentApplicant.universityManual}
+              locked={true}
+              canUnlock={organizer.permissions.includes('super_admin')}
             />
-            <FRQInfo
-              k={'How many hackathons have you submitted to and what did you learn from them?'}
-              v={currentUserGroup[0].hackathonNumber}
-            />
-            <FRQInfo
-              k={
-                "If you haven't been to a hackathon, what do you hope to learn from HackUTD Ripple Effect?"
-              }
-              v={currentUserGroup[0].hackathonFirstTimer}
-            />
-            <FRQInfo
-              k={'What are you looking forward to at HackUTD Ripple Effect?'}
-              v={currentUserGroup[0].lookingForward}
-            />
+            <BasicInfo k="Software Experience" v={currentApplicant.softwareExperience} />
+            {/* this should PROBABLY be currentUser.hackathonNumber, but I think the registration interface is populated incorrectly */}
+            <BasicInfo k="Hackathons Attended" v={`${currentApplicant.hackathonExperience}`} />
           </div>
         </div>
+
+        {/* Application Score */}
+        <div className="flex flex-col items-center sm:w-2/5 text-black">
+          <p className="font-bold text-xl text-black">Application Score</p>
+
+          <p className="text-8xl font-dmSans">
+            {currentApplicant.applicationScore.acceptCount -
+              currentApplicant.applicationScore.rejectCount}
+          </p>
+
+          <p className="italic text-gray-600">
+            <span className="text-green-500">
+              {currentApplicant.applicationScore.acceptCount} accepted
+            </span>{' '}
+            /{' '}
+            <span className="text-red-500">
+              {currentApplicant.applicationScore.rejectCount} rejected
+            </span>
+          </p>
+
+          {user.permissions.includes('super_admin') && (
+            <div className="flex-wrap flex flex-row gap-2 p-3 w-full">
+              <select
+                value={newRole}
+                onChange={(e) => {
+                  setNewRole(e.target.value);
+                }}
+                name="new_role"
+                className="flex-1 border-2 rounded-xl p-2"
+              >
+                <option value="">Choose a role</option>
+                <option value="super_admin">Super Admin</option>
+                <option value="admin">Admin</option>
+                <option value="hacker">Hacker</option>
+              </select>
+
+              {newRole !== '' && (
+                <button
+                  onClick={() => {
+                    updateRole();
+                  }}
+                  className="flex-1 font-bold bg-[rgba(66,184,187,1)] text-white p-2 rounded-xl"
+                >
+                  Update Role
+                </button>
+              )}
+            </div>
+          )}
+        </div>
+      </div>
+
+      {/* Notes */}
+      <div>
+        <textarea
+          ref={applicationNotesRef}
+          className="w-full h-52 border-gray-300 border-2 rounded-md text-black bg-gray-300/20"
+        ></textarea>
+      </div>
+      <div className="my-6 w-full border-2 border-gray-200 rounded-md" />
+      {/* FRQ */}
+      <div className="flex flex-col w-full">
+        <FRQInfo
+          k={'Why do you want to attend HackUTD Ripple Effect?'}
+          v={currentApplicant.whyAttend}
+        />
+        <FRQInfo
+          k={'How many hackathons have you submitted to and what did you learn from them?'}
+          v={currentApplicant.hackathonNumber}
+        />
+        <FRQInfo
+          k={
+            "If you haven't been to a hackathon, what do you hope to learn from HackUTD Ripple Effect?"
+          }
+          v={currentApplicant.hackathonFirstTimer}
+        />
+        <FRQInfo
+          k={'What are you looking forward to at HackUTD Ripple Effect?'}
+          v={currentApplicant.lookingForward}
+        />
       </div>
     </div>
   );
