+++ package.json
@@ -36,6 +36,7 @@
     "commitizen": "^4.2.4",
     "date-fns": "^2.28.0",
     "doppler": "^1.0.0",
+    "embla-carousel-react": "^8.3.0",
     "firebase": "^10.7.1",
     "firebase-admin": "^12.0.0",
     "formik": "^2.2.9",
